migrate TestTallyNoOneVotes to use simapp
