--- x/gov/keeper/old_querier_test.go
@@ -1,1 +1,0 @@
-package keeper
